fix: migrate to aem.js
