+++ blocks/footer/footer.stories.js
@@ -5,7 +5,7 @@ import {
   buildBlock,
   loadBlock,
   decorateBlock,
-} from '../../scripts/lib-franklin.js';
+} from '../../scripts/aem.js';
 import _style from './footer.css';
 import { createTag } from '../../scripts/scripts.js';
