+++ docs/developers/buttons.mdx
@@ -69,7 +69,7 @@ To include an icon in a button you can use the `decorateIcons` function to decor
 The HTMLElement above should then be passed into the `decorateIcons` function to load the icon inside the span. Below is a full example.
 
 ```javascript
-import { decorateIcons } from 'lib-franklin.js';
+import { decorateIcons } from 'scripts.js';
 
 const container = document.createElement('p');
 container.classList.add('button-container');
