+++ docs/stories/components/buttons.stories.js
@@ -2,8 +2,7 @@
 import { within, waitFor, userEvent } from '@storybook/testing-library';
 import { expect } from '@storybook/jest';
 import { FranklinTemplate } from '@dylandepass/franklin-storybook-addon';
-import { decorateIcons } from '../../../scripts/lib-franklin.js';
-import { createTag, loadPage } from '../../../scripts/scripts.js';
+import { createTag, loadPage, decorateIcons } from '../../../scripts/scripts.js';
 
 function decorate(block) {
   decorateIcons(block);
