+++ head.html
@@ -1,5 +1,5 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/scripts/lib-franklin.js" type="module"></script>
+<script src="/scripts/aem.js" type="module"></script>
 <script src="/scripts/scripts.js" type="module"></script>
 <link id="theme" rel="stylesheet" href="/styles/themes/rosalind-light.css" />
 <link rel="stylesheet" href="/styles/system.css">
